(#14) Jump to record/marco after '#' or '?'
vim-erlang-tags.erl generates extra tags (#rec for records), and when C-] is
hit on a record, the # is counted as part of the name, so only records tags
will be considered. If there are several records with the same name, it is
still not guaranteed that C-] will jump to the correct one.

Same was implemented for macros and '?'.