Fix mouse mappings
The original mouse mapping didn't work well, because when I clicked on a word,
Vim thought that the tag was only the part of the word from the position of the
mouse pointer to the end of the word. E.g. if I clicked on the letter "c" of
"function", it wanted to jump to the "ction" tag.