Make autoupdate for single file
Autoupdate consists of three steps. First, using sed (and matching on
the OS so that to understand GNU-sed and BSD-sed), we remove from the
tags file all the tags directing to the current file. Then, we recreate
tags for the current buffer on a temptags file, and then, we merge this
new temptags with the old main tags file, taking care of sorts, using
cmdline sort.