Massive parallelisation of file tagging
Now per each file given, a process is launched to do the work for this
file alone. This way, not only the files are processed in parallel
faster, but we also ensure releasing of the memory that the binaries
take faster, as when the process dies, the ref_count goes down to zero
and is directly collected. Over big big repos, I could measure a drop in
memory usage from 9GB of RAM to 2.5GB!

Note that for this, it is important to create the ets table with
write_concurrency.

An important note is that the functionality to read files from `stdin`
has been removed.