(#46) Lower priority of files in `_build`
Previously the tags coming from the `_build` directory took precedence
over tags coming from the `apps` directory (because the `_` character
precedes the `a` character in ASCII).

Now the tags coming from the `_build` directory have lower priority than
other tags.

As a side effect, this commit modifies the sorting of tags. Previously
the binary lines generated from the tags were sorted. Now the internal
representation ({{TagName, FilePath, Scope, Kind}, TagAddress}) of the
tags are sorted instead. This made it necessary to store TagName,
FilePath and TagAddress as binaries instead of iolists. (Because iolists
are not sorted the way we want. E.g., [["a"]] is larger than ["b"].)

Authors:

- haoxian <haoxianhan@gmail.com>
- Csaba Hoch <csaba.hoch@gmail.com>