Jump to `f(` instead of `f\>`
Previously vim-erlang-tags generated the following lines for jumping to
function definitions ("f" is the function name):

    mymod:f <tab> ./mymod.erl <tab> /^f\>/
    f <tab> ./mymod.erl <tab> /^f\>/ <space><space> ;" <tab> file:

Erlang/OTP 27 introduced docstrings, which means that now we can have
lines such as these:

    -doc """
    f
    """.

The /^f\>/ pattern matches the second line, so if this docstring
precedes the definition of the `f` function, Vim will jump to the "f" in
the docstring instead of jumping to the `f` function definition.

After this commit, the pattern is more strict:

    mymod:f <tab> ./mymod.erl <tab> /^f(/
    f <tab> ./mymod.erl <tab> /^f(/ <space><space> ;" <tab> file:

So Vim will jump to the function definition instead of the docstring
above.

The new pattern is still not perfect because it can be fooled by the
following:

    -doc """
    f(
    """.

Also, if someone writes a space character or a newline between the
function name and the opening parenthesis, the pattern won't find the
function definition:

    f ()